Fix for RCE on GetText.php
Added regular expression match for number as beginning part of $number and rest is ignored. This is helpful is user enter numbers. Such as 123asd, 456%$# will be matched and accepted as 123 and 456. $number will be set to blank if not matched. Developer can decide at this point what to do when not matched.